Rectangle can calculate area and perimeter with double values

+++ src/com/tw/step/assignment1/shapes/Rectangle.java
@@ -1,19 +1,19 @@
 package com.tw.step.assignment1.shapes;
 
 public class Rectangle {
-  private final int length;
-  private final int breadth;
+  private final double length;
+  private final double breadth;
 
-  public Rectangle(int length, int breadth) {
+  public Rectangle(double length, double breadth) {
     this.length = length;
     this.breadth = breadth;
   }
 
-  public int area() {
+  public double area() {
     return length * breadth;
   }
 
-  public int perimeter() {
+  public double perimeter() {
     return 2*(this.length+this.breadth);
   }
 }
